\name{MitISEM-package}
\alias{MitISEM-package}
\docType{package}
\title{Mixture of Student t Distributions using Importance Sampling and Expectation Maximization
}
\description{Appeoximates a univariate or multivariate density using mixture of Student \code{t} distributions, achieved by Importance Sampling and Expectation Maximization algorithms
}
\details{
\tabular{ll}{
Package: \tab MitISEM\cr
Type: \tab Package\cr
Version: \tab 1.1\cr
Date: \tab 2017-04-25\cr
License: \tab GPL (>= 3)\cr
}
Flexible multivariate function approximation using adapted
Mixture of Student t Distributions. Mixture of t distribution
is obtained using Importance Sampling weighted Expectation
Maximization algorithm.
}
\author{N. Basturk, L.F. Hoogerheide, A. Opschoor, H.K. van Dijk
}
\references{L. Hoogerheide, A. Opschoor and H.K. Van Dijk (2012). "A class of adaptive importance sampling weighted EM algorithms for efficient and robust posterior and predictive simulation." Journal of Econometrics 171(2): 101-120.
}
\keyword{ package }
