\name{MitISEM-package}
\alias{MitISEM-package}
\docType{package}
\title{
Mixture of Student t Distributions using Importance Sampling and Expectation Maximization
}
\description{
Approximates a univariate or multivariate density using mixture of Student \code{t} distributions, achieved by Importance Sampling and Expectation Maximization algorithms
}
\details{

\tabular{ll}{
Package: \tab MitISEM\cr
Type: \tab Package\cr
Version: \tab 1.2\cr
Date: \tab 2017-07-10\cr
License: \tab GPL (>= 3)\cr
}
Flexible multivariate function approximation using adapted
Mixture of Student t Distributions. Mixture of t distribution
is obtained using Importance Sampling weighted Expectation
Maximization algorithm.
}
\author{
N. Basturk, L.F. Hoogerheide, A. Opschoor, H.K. van Dijk

Maintainer: N. Basturk
}
\references{
Basturk, N., Grassi, S., Hoogerheide, L., Opschoor, A. and Van Dijk, H. K. (2017) 
The R Package MitISEM: Efficient and Robust Simulation Procedures for Bayesian Inference. 
\emph{Journal of Statistical Software}, 79(1): 1-39. \doi{10.18637/jss.v079.i01}.

Hoogerheide L., Opschoor, A. and Van Dijk, H. K. (2012) 
A Class of Adaptive Importance Sampling Weighted {EM} Algorithms for Efficient and Robust Posterior and Predictive Simulation. 
\emph{Journal of Econometrics}, 171(2): 101-120.
\url{http://www.sciencedirect.com/science/article/pii/S0304407612001583}.
}
\keyword{ package }
