\name{MitISEM-package}
\alias{MitISEM-package}
\docType{package}
\title{Mixture of Student t Distributions using Importance Sampling and Expectation Maximization
}
\description{Appeoximates a univariate or multivariate density using mixture of Student \code{t} distributions, achieved by Importance Sampling and Expectation Maximization algorithms
}
\details{
\tabular{ll}{
Package: \tab MitISEM\cr
Type: \tab Package\cr
Version: \tab 1.0\cr
Date: \tab 2012-07-31\cr
License: \tab What license is it under?\cr
}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{N. Basturk, L.F. Hoogerheide, A. Opschoor, H.K. van Dijk
}
\references{'A Class of Adaptive EM-based Importance Sampling Algorithms for Efficient and Robust Posterior and Predictive Simulation', 2012,, by Lennart Hoogerheide and Anne Opschoor and Herman K. van Dijk", Tinbergen Institute Discussion Paper, no. 11-004/4
}
\keyword{ package }
\examples{
  print("to be completed")
}
